maka MessageCache public, added to user
