+++ src/test/java/pl/lcc/listener/example/service/InMemoryMessageServiceTest.java
@@ -4,11 +4,10 @@
  */
 package pl.lcc.listener.example.service;
 
-import java.time.LocalDateTime;
-import java.util.stream.Stream;
+
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import pl.lcc.listener.example.user.Message;
+
 
 /**
  *
@@ -17,32 +16,5 @@ import pl.lcc.listener.example.user.Message;
 public class InMemoryMessageServiceTest {
     
   
-    @Test
-    public void testAddingMessagesToCache() {
-        
-        var sut = new InMemoryMessageService.MessageCache();
-       
-        //fill three elements to list
-        Stream
-                .of("1","2","3")
-                .map(n -> new Message(LocalDateTime.now(), n, "some user"))
-                .forEach(msg-> sut.addMessage(msg));
-        var result = sut.getList();                
-        Assertions.assertThat(result).map(msg -> msg.getMessageBody()).contains("1", "2", "3");
-        
-        //add fourth -> 1st should be removed
-        sut.addMessage(new Message(LocalDateTime.now(), "Other message", "any user"));        
-        var resultAfter4th = sut.getList();
-        Assertions.assertThat(resultAfter4th).map(msg -> msg.getMessageBody()).contains("2", "3", "Other message");
-    }
-    
-    @Test
-    void testEmptyCache(){
-          var sut = new InMemoryMessageService.MessageCache();
-        
-        var emptyResult = sut.getList();
-        System.out.println(emptyResult);        
-        Assertions.assertThat(emptyResult).isEmpty();
-    }
     
 }
